refactor: move the snapcast control to the snapcast package folder
