+++ app/src/main/java/tech/capullo/radio/compose/RadioBroadcasterScreen.kt
@@ -21,12 +21,12 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
-import tech.capullo.radio.data.Client
-import tech.capullo.radio.data.ClientConfig
-import tech.capullo.radio.data.Host
-import tech.capullo.radio.data.LastSeen
-import tech.capullo.radio.data.SnapClient
-import tech.capullo.radio.data.Volume
+import tech.capullo.radio.snapcast.Client
+import tech.capullo.radio.snapcast.ClientConfig
+import tech.capullo.radio.snapcast.Host
+import tech.capullo.radio.snapcast.LastSeen
+import tech.capullo.radio.snapcast.SnapClient
+import tech.capullo.radio.snapcast.Volume
 import tech.capullo.radio.ui.theme.Typography
 import tech.capullo.radio.ui.theme.primaryBlack
 import tech.capullo.radio.ui.theme.secondaryOrange
